moving handlesubmit into the backend folder
