--- frontend/Amortization.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Amortization = () => {
-
-    
-
-}
-
-export default Amortization
